UsersテーブルとListingsテーブルの結合・auth()機能を用いた機能制限の実装

+++ app/Http/Controllers/ListingController.php
@@ -46,6 +46,8 @@ class ListingController extends Controller
             $formFields['logo'] = $request->file('logo')->store('logos', 'public');
         }
 
+        $formFields['user_id'] = auth()->id();
+
         // 送信されたデータをデータベースに保存する
         Listing::create($formFields);
 
@@ -64,6 +66,11 @@ class ListingController extends Controller
 
     // 求人情報を編集・更新する
     public function update(Request $request, Listing $listing) {
+
+        // ログインしているユーザーが、求人の投稿者であることを確認する
+        if($listing->user_id != auth()->id()) {
+            abort(403, 'ユーザー情報を認証できませんでした。');
+        }
 
         // バリデーション
         $formFields = $request->validate([
@@ -89,8 +96,21 @@ class ListingController extends Controller
 
     // 求人を削除する
     public function destroy(Listing $listing) {
+
+        // ログインしているユーザーが、求人の投稿者であることを確認する
+        if($listing->user_id != auth()->id()) {
+            abort(403, 'ユーザー情報を認証できませんでした。');
+        }
+
         $listing->delete();
 
         return redirect('/')->with('message', '求人が削除されました。');
+    }
+
+    // 求人情報管理画面を表示する
+    public function manage() {
+        return view('listings.manage', [
+            'listings' => auth()->user()->listings()->get()
+        ]);
     }
 }
